# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Talven",
    "description": "Talven briefing service",
    "version": "0.20.6"
  },
  "paths": {
...
    "/briefings": {
      "get": {
        "tags": [
          "briefings"
        ],
        "summary": "List Briefings",
        "operationId": "list_briefings_briefings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 24,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "newest",
                "oldest"
              ],
              "type": "string",
              "default": "newest",
              "title": "Sort"
            }
          },
          {
            "name": "sourceType",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "youtube",
                "url"
              ],
              "type": "string",
              "default": "all",
              "title": "Sourcetype"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefingListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/billing/checkout": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Checkout",
        "operationId": "create_checkout_billing_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream provider failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/billing/portal": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Portal",
        "operationId": "create_portal_billing_portal_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerPortalSessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream provider failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/billing/packs/{polar_order_id}/refund": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Refund Pack",
        "operationId": "refund_pack_billing_packs__polar_order_id__refund_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "polar_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Polar Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackRefundResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream provider failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/billing/operations/{operation_id}": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get Operation",
        "operationId": "get_operation_billing_operations__operation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Operation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingSyncOperationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Billing operation not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/billing/plans": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "List Plans",
        "operationId": "list_plans_billing_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanResponse"
                  },
                  "type": "array",
                  "title": "Response List Plans Billing Plans Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream provider failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/billing/usage": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get Usage",
        "operationId": "get_usage_billing_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageOverviewResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/billing/briefings": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get Briefings",
        "operationId": "get_briefings_billing_briefings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsageHistoryEntry"
                  },
                  "type": "array",
                  "title": "Response Get Briefings Billing Briefings Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/billing/account": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get Account",
        "operationId": "get_account_billing_account_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccountResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BillingSyncOperationResponse": {
        "properties": {
          "operation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Operation Id"
          },
          "operation_type": {
            "type": "string",
            "enum": [
              "checkout",
              "refund"
            ],
            "title": "Operation Type"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "succeeded",
              "failed"
            ],
            "title": "Status"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          }
        },
        "type": "object",
        "required": [
          "operation_id",
          "operation_type",
          "status",
          "failure_code"
        ],
        "title": "BillingSyncOperationResponse"
      },
...
      "BriefingListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BriefingListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "sort": {
            "type": "string",
            "enum": [
              "newest",
              "oldest"
            ],
            "title": "Sort"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "all",
              "youtube",
              "url"
            ],
            "title": "Source Type"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total_count",
          "limit",
          "offset",
          "has_more",
          "sort",
          "source_type"
        ],
        "title": "BriefingListResponse"
      },
...
      "CheckoutSessionRequest": {
        "properties": {
          "plan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Plan Id"
          }
        },
        "type": "object",
        "required": [
          "plan_id"
        ],
        "title": "CheckoutSessionRequest"
      },
      "CheckoutSessionResponse": {
        "properties": {
          "checkout_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Checkout Url"
          },
          "operation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Operation Id"
          }
        },
        "type": "object",
        "required": [
          "checkout_url",
          "operation_id"
        ],
        "title": "CheckoutSessionResponse"
      },
      "CustomerPortalSessionResponse": {
        "properties": {
          "portal_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Portal Url"
          }
        },
        "type": "object",
        "required": [
          "portal_url"
        ],
        "title": "CustomerPortalSessionResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PackRefundResponse": {
        "properties": {
          "operation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Operation Id"
          },
          "polar_order_id": {
            "type": "string",
            "title": "Polar Order Id"
          },
          "refund_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refund Id"
          },
          "requested_amount_cents": {
            "type": "integer",
            "title": "Requested Amount Cents"
          },
          "remaining_seconds_before_refund": {
            "type": "integer",
            "title": "Remaining Seconds Before Refund"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "operation_id",
          "polar_order_id",
          "refund_id",
          "requested_amount_cents",
          "remaining_seconds_before_refund",
          "status"
        ],
        "title": "PackRefundResponse"
      },
      "PlanResponse": {
        "properties": {
          "plan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Plan Id"
          },
          "plan_code": {
            "type": "string",
            "title": "Plan Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "plan_type": {
            "type": "string",
            "title": "Plan Type"
          },
          "polar_product_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polar Product Id"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "amount_cents": {
            "type": "integer",
            "title": "Amount Cents"
          },
          "billing_interval": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Interval"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "quota_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota Seconds"
          },
          "rollover_cap_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rollover Cap Seconds"
          },
          "pack_expiry_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Expiry Days"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "plan_code",
          "name",
          "plan_type",
          "polar_product_id",
          "currency",
          "amount_cents",
          "billing_interval",
          "version",
          "quota_seconds",
          "rollover_cap_seconds",
          "pack_expiry_days",
          "is_active"
        ],
        "title": "PlanResponse"
      },
...
      "UsageHistoryEntry": {
        "properties": {
          "job_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "seconds_used": {
            "type": "integer",
            "title": "Seconds Used"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "session_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Path"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "seconds_used",
          "source",
          "created_at"
        ],
        "title": "UsageHistoryEntry"
      },
      "UsageOverviewResponse": {
        "properties": {
          "subscription_plan_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Plan Name"
          },
          "has_active_paid_subscription": {
            "type": "boolean",
            "title": "Has Active Paid Subscription"
          },
          "subscription_remaining_seconds": {
            "type": "integer",
            "title": "Subscription Remaining Seconds"
          },
          "pack_remaining_seconds": {
            "type": "integer",
            "title": "Pack Remaining Seconds"
          },
          "total_remaining_seconds": {
            "type": "integer",
            "title": "Total Remaining Seconds"
          },
          "pack_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Expires At"
          },
          "debt_seconds": {
            "type": "integer",
            "title": "Debt Seconds"
          },
          "is_blocked": {
            "type": "boolean",
            "title": "Is Blocked"
          }
        },
        "type": "object",
        "required": [
          "subscription_plan_name",
          "has_active_paid_subscription",
          "subscription_remaining_seconds",
          "pack_remaining_seconds",
          "total_remaining_seconds",
          "pack_expires_at",
          "debt_seconds",
          "is_blocked"
        ],
        "title": "UsageOverviewResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}